Rename spack.util.spack_filesystem -> spack.util.filesystem
