+++ lib/spack/spack/util/filesystem.py
